Fix permission list validation for root node.
When root node is updated and does not contain `:config` then the
broker will inject `:config` into the parameters for next time but
did not inject into the permissions currently being injected.
Change takes advantage of Dart's pass-by-reference to inject the
config permissions into the list passed in, if it was not found.